requirements.txt
fastapi
uvicorn
python-dotenv
alpaca-py
pandas
